fastapi code decoupling
